+++ seedemu/__init__.py
@@ -7,4 +7,3 @@ from .mergers import *
 from .components import *
 from .raps import *
 from .utilities import *
-from .docker import *
