+++ seedemu/compiler/docker_images/EthereumImage.py
@@ -1,18 +1,18 @@
-from DockerPreBuiltImage import DockerPreBuiltImage
-from seedemu.docker.BaseImage import BaseImage
+from .DockerPreBuiltImage import DockerPreBuiltImage
+from .BaseImage import BaseImage
 
-class RouterImage(DockerPreBuiltImage):
+class EthereumImage(DockerPreBuiltImage):
     @property
     def _base_image(self):
-        return "handsonsecurity/seedemu-router"
+        return "handsonsecurity/seedemu-ethereum"
 
     @property
     def _installedSoftware(self) -> set:
         """!
         brief get the installed software if using a pre-built image.
         """
-        return set({"bird2"})
-    
+        return set({})
+
     @property
     def _subset(self) -> DockerPreBuiltImage:
         """!
