+++ seedemu/compiler/docker_images/RouterImage.py
@@ -1,18 +1,18 @@
-from DockerPreBuiltImage import DockerPreBuiltImage
-from seedemu.docker.BaseImage import BaseImage
+from .DockerPreBuiltImage import DockerPreBuiltImage
+from .BaseImage import BaseImage
 
-class EthereumImage(DockerPreBuiltImage):
+class RouterImage(DockerPreBuiltImage):
     @property
     def _base_image(self):
-        return "handsonsecurity/seedemu-ethereum"
+        return "handsonsecurity/seedemu-router"
 
     @property
     def _installedSoftware(self) -> set:
         """!
         brief get the installed software if using a pre-built image.
         """
-        return set({})
-
+        return set({"bird2"})
+    
     @property
     def _subset(self) -> DockerPreBuiltImage:
         """!
